requirements.txt
appnope==0.1.0
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
ConfigArgParse==1.2.3
decorator==4.4.2
Flask==1.1.2
Flask-BasicAuth==0.2.0
gevent==20.6.2
geventhttpclient==1.4.4
greenlet==0.4.16
grpcio==1.31.0
grpcio-tools==1.31.0
idna==2.10
ipdb==0.13.3
ipython==7.18.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
locust==1.2.3
MarkupSafe==1.1.1
msgpack==1.0.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
protobuf==3.13.0
psutil==5.7.2
ptyprocess==0.6.0
Pygments==2.6.1
pyzmq==19.0.2
requests==2.24.0
six==1.15.0
traitlets==5.0.2
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
zope.event==4.4
zope.interface==5.1.0
